Move (un)hex/bin functions to separate file
